Rethink debug printouts.
Most debug printouts were previously out-commented, and to enable them
we had to de-comment the printouts, fixing any new compilation issues
that arose because of non-maintenance of out-commented code.

We now introduce a better approach. Debug printouts can be
enabled/disabled per-package, and even per-class at runtime by setting
the value of a system property to ”true” or ”false”.