IOLog now dynamically determines the class + method name of the caller.
We no longer need to supply the class name to IOLog on creation and
method name when we need to print it out. Instead we determine this
through the stack trace retrieved from Thread.getStackTrace().
This is much less error-prone than the previous "manual" approach,
although it does come with some performance penalty (I miss C's
no-overhead compile-time __FILE__, __LINE__ and __FUNCTION__ macros).