SynchronizedReadableRancomAccessStream.java: Lazily close source stream.
Previously the approach to closing the source stream was that the caller
should always make sure to close all substreams that are using this
SynchronizedReadableRandomAccessStream before calling "close()", which
unconditionally closes the source stream. If the caller fails to do
that, and exception is thrown and the close does not succeed.

However, this was found to be an unpractical approach since the caller
may not always have control over which substreams are created and when,
so instead we introduce the following scheme:
- On creating of the SynchronizedReadableRandomAccessStream, a reference
  is taken.
- Each new substream takes a reference, but after all substreams are
  closed there is still a reference, i.e. the "self"-reference.
- Only when the number of references reaches 0 is the source stream
  closed. This can happen on close() or it can happen on the last
  dereferencing by a substream.

This means that close() no longer guarantees that the underlying stream
is closed... it is closed only if the last substream has released its
reference. If this has not happened at the time that "close()" is
called, then the underlying stream will be kept open until the last
substream releases its reference.