SynchronizedReadableRandomAccessStream.java: Prevent bad dereferences.
If the substream would call "removeReference(Object)" multiple times, we
cannot detect it but we can at least prevent refCount from decreasing
beyond a specified point. If the stream is open, we must at least have a
reference count of 1, and otherwise at least a reference count of 0
(a negative reference count doesn't make sense).