BasicConcatenatedStream.java: Fix reading beyond bounds of stream.
The logic in BasicConcatenatedStream.read(byte[], int, int) to skip to
the part containing the first byte of the read was flawed if the file
pointer pointed at the final byte of the concatenated stream, in which
case we could just keep reading beyond its bounds, advancing the file
pointer far beyond the defined length of the stream.

Fixed by properly detecting when we are reading beyond the last part and
returning -1 if we do.