Log.java: Make log methods final to aid inlining.
We want the Log methods to be inlined as often as possible for
performance reasons and Java is more keen to inline methods when they
are final and immutable.
It's unclear how much this really helps with modern JVMs but it's a less
complex case for the JIT to process so it may still have some benefit.